# document/openapi/openapi-auth.yaml
openapi: 3.0.3
info:
  title: Gather Auth API
  description: |
    Gather 서비스 인증 및 회원가입 관련 API 명세입니다.

    프론트 연동 참고사항:
    - `signup`, `login`, `logout`, `reissue`는 인증 행위 API이므로 동사형 경로를 사용합니다.
    - `regions`, `categories` 조회 API는 auth 도메인 소유는 아니지만 회원가입 화면에서 필요하므로 함께 명세에 포함했습니다.
    - 회원가입 요청에는 관심 카테고리 enum 값 배열인 `interestCategories`를 전달합니다. 값은 `PostingCategory`(ENVIRONMENT, EDUCATION, CULTURE, COMMUNITY, WELFARE, OVERSEAS) 중 중복 없이 1개 이상이며, 정의되지 않은 값은 400(`VALIDATION_ERROR`)입니다.
    - `GET /api/v1/categories`는 아직 기존 16종 분류(`id`/`code`/`name`)를 반환합니다. 6종 enum 전환은 별도 작업으로 진행 중이므로, 전환 완료 전까지 회원가입 카테고리 선택지는 프론트에서 6종 enum 기준으로 처리합니다.
    - Refresh Token은 응답 body가 아니라 `gather_refresh_token` HttpOnly 쿠키로 전달합니다.
    - 재발급과 로그아웃은 Access Token 없이 Refresh Token 쿠키만으로 요청합니다.
    - 로그아웃은 같은 Refresh Token 쿠키로 여러 번 요청해도 성공 응답을 반환합니다.
    - 휴대폰 인증은 Gather가 5분 유효한 인증문구를 만들고, 사용자가 OCTOMO 대표번호로 직접 문자를 보낸 뒤 confirm API로 확인하는 번호 점유 인증입니다. 인증 완료 결과는 30분 안에 해당 `verificationId`와 전화번호를 함께 제출한 일반·카카오 회원가입에서 한 번만 소비할 수 있습니다.
    - offset 없는 이메일 인증 시각(`expiresAt`, `resendAvailableAt`, `verifiedAt`)은 UTC 기준입니다. 휴대폰 인증 `expiresAt`은 UTC offset(`Z`)을 포함합니다.
    - 카카오 로그인은 `POST /api/v1/auth/kakao/login`에 `{authorizationCode, redirectUri}`를 전달합니다. `redirectUri`는 서버 허용 목록과 문자열까지 정확히 일치해야 합니다. 응답은 200이며 `signupStatus`로 기존 회원(`LOGIN_COMPLETED`)과 신규 회원(`ADDITIONAL_INFO_REQUIRED`)을 구분합니다. `ADDITIONAL_INFO_REQUIRED`는 에러가 아닙니다.
...
  version: 1.0.0
servers:
  - url: /
    description: Relative server URL
...
paths:
...
  /api/v1/auth/phone-verifications:
    post:
      tags:
        - Auth
      summary: 휴대폰 문자 인증 시작
      description: 010 휴대폰 번호를 정규화하고 5분 유효한 인증 세션을 만듭니다. 같은 번호의 연속 시작 요청은 60초 동안 제한합니다. 모바일은 응답의 receiverNumber와 messageText로 sms URI를 구성합니다.
      operationId: startPhoneVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneVerificationStartRequest'
            example:
              phoneNumber: '01012345678'
      responses:
        '201':
          description: 휴대폰 인증 세션 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneVerificationStartApiResponse'
              example:
                success: true
                data:
                  verificationId: 5c5d5db1-4187-43d0-8580-672307994878
                  receiverNumber: '16663538'
                  messageText: GATHER-7F2K9Q8M4P
                  expiresAt: '2026-08-09T06:50:00Z'
                error: null
        '400':
          description: 010으로 시작하는 11자리 휴대폰 번호가 아님
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: VALIDATION_ERROR
                  message: 요청 값이 올바르지 않습니다.
        '429':
          description: 같은 번호의 인증 시작 요청 간격 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_RATE_LIMITED
                  message: 휴대폰 인증 요청이 많습니다. 잠시 후 다시 시도해주세요.
  /api/v1/auth/phone-verifications/{verificationId}/qr-code:
    post:
      tags:
        - Auth
      summary: PC용 휴대폰 인증 QR 발급
      description: request body를 받지 않고 서버에 저장된 인증문구로 OCTOMO SMS QR data URL을 발급합니다. 같은 인증 세션은 10초 간격, 최대 3회로 제한합니다.
      operationId: createPhoneVerificationQrCode
      parameters:
        - $ref: '#/components/parameters/PhoneVerificationId'
      responses:
        '201':
          description: QR 발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneVerificationQrCodeApiResponse'
              example:
                success: true
                data:
                  qrCode: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
                error: null
        '400':
          description: verificationId 형식 오류 또는 인증 세션 만료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                validationError:
                  summary: verificationId 형식 오류
                  value:
                    success: false
                    data: null
                    error:
                      code: VALIDATION_ERROR
                      message: 요청 값이 올바르지 않습니다.
                phoneVerificationExpired:
                  summary: 인증 세션 만료
                  value:
                    success: false
                    data: null
                    error:
                      code: PHONE_VERIFICATION_EXPIRED
                      message: 휴대폰 인증 요청이 만료되었습니다.
        '404':
          description: 인증 세션 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_NOT_FOUND
                  message: 휴대폰 인증 요청을 찾을 수 없습니다.
        '429':
          description: QR 발급 횟수·간격 제한 또는 OCTOMO 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_RATE_LIMITED
                  message: 휴대폰 인증 요청이 많습니다. 잠시 후 다시 시도해주세요.
        '503':
          description: OCTOMO 설정·네트워크·서비스 장애
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_PROVIDER_UNAVAILABLE
                  message: 휴대폰 인증 서비스를 일시적으로 사용할 수 없습니다.
  /api/v1/auth/phone-verifications/{verificationId}/confirm:
    post:
      tags:
        - Auth
      summary: 휴대폰 문자 인증 확인
      description: request body를 받지 않고 서버 저장 전화번호와 인증문구를 OCTOMO에서 확인합니다. 문자가 아직 없으면 오류가 아닌 PENDING을 반환합니다. 같은 인증 세션은 3초 간격, 최대 30회로 제한합니다.
      operationId: confirmPhoneVerification
      parameters:
        - $ref: '#/components/parameters/PhoneVerificationId'
      responses:
        '200':
          description: 문자 수신 대기 또는 휴대폰 인증 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneVerificationConfirmApiResponse'
              examples:
                pending:
                  value:
                    success: true
                    data:
                      status: PENDING
                    error: null
                verified:
                  value:
                    success: true
                    data:
                      status: VERIFIED
                    error: null
        '400':
          description: verificationId 형식 오류 또는 인증 세션 만료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                validationError:
                  summary: verificationId 형식 오류
                  value:
                    success: false
                    data: null
                    error:
                      code: VALIDATION_ERROR
                      message: 요청 값이 올바르지 않습니다.
                phoneVerificationExpired:
                  summary: 인증 세션 만료
                  value:
                    success: false
                    data: null
                    error:
                      code: PHONE_VERIFICATION_EXPIRED
                      message: 휴대폰 인증 요청이 만료되었습니다.
        '404':
          description: 인증 세션 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_NOT_FOUND
                  message: 휴대폰 인증 요청을 찾을 수 없습니다.
        '409':
          description: 이미 가입된 전화번호 또는 탈퇴 후 재가입 제한 중인 전화번호
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                duplicatePhoneNumber:
                  summary: 이미 가입된 전화번호
                  value:
                    success: false
                    data: null
                    error:
                      code: DUPLICATE_PHONE_NUMBER
                      message: 이미 사용 중인 전화번호입니다.
                accountRejoinBlocked:
                  summary: 탈퇴 후 재가입 제한 중인 전화번호
                  value:
                    success: false
                    data: null
                    error:
                      code: ACCOUNT_REJOIN_BLOCKED
                      message: 탈퇴 후 7일 동안 재가입할 수 없습니다.
        '429':
          description: 확인 횟수·간격 제한 또는 OCTOMO 요청 제한
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_RATE_LIMITED
                  message: 휴대폰 인증 요청이 많습니다. 잠시 후 다시 시도해주세요.
        '503':
          description: OCTOMO 설정·네트워크·서비스 장애
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: PHONE_VERIFICATION_PROVIDER_UNAVAILABLE
                  message: 휴대폰 인증 서비스를 일시적으로 사용할 수 없습니다.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PhoneVerificationId:
      name: verificationId
      in: path
      required: true
      description: 휴대폰 인증 세션 식별자
      schema:
        type: string
        format: uuid
  schemas:
...
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 에러 코드
          example: VALIDATION_ERROR
        message:
          type: string
          description: 에러 메시지
          example: 요청 값이 올바르지 않습니다.
    ErrorApiResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ApiError'
...
    PhoneVerificationStartRequest:
      type: object
      required:
        - phoneNumber
      properties:
        phoneNumber:
          type: string
          maxLength: 20
          description: 인증할 전화번호. 하이픈과 공백은 서버에서 제거하며 정규화 결과는 010으로 시작하는 11자리 숫자여야 합니다.
          example: '01012345678'
    PhoneVerificationStartData:
      type: object
      required:
        - verificationId
        - receiverNumber
        - messageText
        - expiresAt
      properties:
        verificationId:
          type: string
          format: uuid
        receiverNumber:
          type: string
          example: '16663538'
        messageText:
          type: string
          description: 모바일 SMS body와 OCTOMO 조회에 그대로 사용할 인증코드
          example: GATHER-7F2K9Q8M4P
        expiresAt:
          type: string
          format: date-time
          description: 인증 세션 만료 시각(UTC)
          example: '2026-08-09T06:50:00Z'
    PhoneVerificationStartApiResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PhoneVerificationStartData'
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ApiError'
          example: null
    PhoneVerificationQrCodeData:
      type: object
      required:
        - qrCode
      properties:
        qrCode:
          type: string
          description: img src에 바로 사용할 수 있는 PNG data URL
          example: data:image/png;base64,iVBORw0KGgo...
    PhoneVerificationQrCodeApiResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PhoneVerificationQrCodeData'
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ApiError'
          example: null
    PhoneVerificationConfirmData:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - PENDING
            - VERIFIED
    PhoneVerificationConfirmApiResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PhoneVerificationConfirmData'
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ApiError'
          example: null
...